refactor: separate vm.rs into 2 different modules
levm and revm modules
