+++ crates/vm/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 ethrex-core = { path = "../common", default-features = false }
 ethrex-storage = { path = "../storage/store", default-features = false }
-ethrex-levm = { path = "./levm", optional = true }
+ethrex-levm = { path = "./levm" }
 ethrex-trie = { path = "../storage/trie", default-features = false }
 ethrex-rlp = { path = "../common/rlp", default-features = false }
 revm = { version = "14.0.3", features = [
@@ -39,7 +39,7 @@ path = "./vm.rs"
 
 [features]
 default = ["c-kzg", "blst"]
-levm = ["default", "ethrex-levm"]
+levm = ["default"]
 l2 = []
 c-kzg = ["revm/c-kzg"]
 blst = ["revm/blst"]
